covert decimal to double

+++ KLD/Clusterer/Clusterer/IDistance.cs
@@ -8,6 +8,6 @@ namespace Clusterer
 {
     public interface IDistance
     {
-        decimal GetDistance(Document di, Document dj);
+        double GetDistance(Document di, Document dj);
     }
 }
